backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["coffree <02320013apis@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
pre-commit = "^3.5.0"
fastapi = "^0.104.0"
uvicorn = "^0.23.2"
pylint = "^3.0.2"
psycopg2-binary = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
